+++ src/main/kotlin/racingcar/dto/round/RoundDto.kt
@@ -1,0 +1,3 @@
+package racingcar.dto.round
+
+class RoundDto(val count: Int)
