+++ src/main/kotlin/racingcar/service/RacingService.kt
@@ -2,7 +2,7 @@ package racingcar.service
 
 import racingcar.dto.car.CarsDto
 import racingcar.dto.car.WinnersDto
-import racingcar.model.round.RoundDto
+import racingcar.dto.round.RoundDto
 import racingcar.utils.mapper.toDto
 import racingcar.utils.mapper.toModel
 import racingcar.utils.random.MovementProbabilityGenerator
