+++ src/main/kotlin/racingcar/utils/mapper/RoundMapper.kt
@@ -1,0 +1,6 @@
+package racingcar.utils.mapper
+
+import racingcar.dto.round.RoundDto
+import racingcar.model.round.Round
+
+fun RoundDto.toModel(): Round = Round(count)
